Move `<CardDivider />` to `card-divider/` subfolder
