+++ packages/components/src/card/index.js
@@ -1,2 +1,3 @@
 export { default as Card } from './card';
 export { default as CardBody } from './card-body';
+export { default as CardDivider } from './card-divider';
