+++ packages/components/src/index.js
@@ -25,8 +25,7 @@ export { default as BaseControl } from './base-control';
 export { default as __experimentalBoxControl } from './box-control';
 export { default as Button } from './button';
 export { default as ButtonGroup } from './button-group';
-export { Card, CardBody } from './card';
-export { default as CardDivider } from './card/divider';
+export { Card, CardBody, CardDivider } from './card';
 export { default as CardFooter } from './card/footer';
 export { default as CardHeader } from './card/header';
 export { default as CardMedia } from './card/media';
